requirements.txt
psycopg2-binary~=2.9.10
PyYAML~=6.0.2
h3~=4.2.2
selenium~=4.30.0
yandex-geocoder==3.0.1
requests~=2.32.3
beautifulsoup4~=4.13.4
tqdm~=4.67.1
jsonpickle~=4.0.5